--- a/app/components/courses/summary_card_component.rb
+++ b/app/components/courses/summary_card_component.rb
@@ -95,8 +95,7 @@ module Courses
       t('.length_key')
     end
 
-    def length_value
-      course_length = course.enrichment_attribute(:course_length).to_s
+    def length_value(course_length = enrichment.course_length)
       translated_course_length = t(".length_value.#{course_length}", default: course_length)
 
       [translated_course_length, course.study_mode.humanize.downcase].join(' - ')
@@ -154,11 +153,11 @@ module Courses
       t(".location_value.school_term.#{course.funding}", default: t('.location_value.school_term.default'))
     end
 
-    def uk_fees(fee_uk = course.enrichment_attribute(:fee_uk_eu))
+    def uk_fees(fee_uk = enrichment.fee_uk_eu)
       t('.fee_value.fee.uk_fees_html', value: content_tag(:b, number_to_currency(fee_uk.to_f)))
     end
 
-    def international_fees(fee_international = course.enrichment_attribute(:fee_international))
+    def international_fees(fee_international = enrichment.fee_international)
       return if fee_international.blank?
 
       t('.fee_value.fee.international_fees_html', value: content_tag(:b, number_to_currency(fee_international.to_f)))
@@ -208,7 +207,15 @@ module Courses
     end
 
     def main_subject
-      @main_subject ||= Subject.find_by(id: course.master_subject_id)
+      @main_subject ||= course.subjects.find { |subject| subject.id == course.master_subject_id }
+    end
+
+    # rubocop:disable Lint/UselessConstantScoping
+    NullEnrichment = Struct.new(:course_length, :fee_uk_eu, :fee_international, keyword_init: true)
+    # rubocop:enable Lint/UselessConstantScoping
+
+    def enrichment
+      @enrichment ||= course.latest_published_enrichment || NullEnrichment.new
     end
 
     def bursary_and_scholarship_flag_active_or_preview?