--- a/app/models/course.rb
+++ b/app/models/course.rb
@@ -27,17 +27,6 @@ class Course < ApplicationRecord
   audited
 
   attribute :subordinate_subject_id, :integer
-  after_initialize do
-    if is_primary?
-      self.master_subject_id ||= course_subjects.first&.subject_id
-    else
-      sub = course_subjects.select(&:position)
-
-      self.master_subject_id ||= sub.first&.subject&.id
-
-      self.subordinate_subject_id ||= sub&.second&.subject&.id
-    end
-  end
 
   validates :course_code,
             uniqueness: { scope: :provider_id },
@@ -204,7 +193,7 @@ class Course < ApplicationRecord
     end
   end
 
-  has_one :latest_published_enrichment, -> { published.order('created_at DESC, id DESC').limit(1) },
+  has_one :latest_published_enrichment, -> { published.most_recent },
           class_name: 'CourseEnrichment', inverse_of: :course
 
   scope :within, lambda { |range, origin:|
@@ -905,32 +894,8 @@ class Course < ApplicationRecord
     "#{name} (#{course_code})"
   end
 
-  def available_placements
-    @available_placements ||= site_statuses
-                              .joins(:site)
-                              .where(
-                                status: [SiteStatus.statuses[:running], SiteStatus.statuses[:new_status]],
-                                publish: SiteStatus.publishes[:published],
-                                vac_status: [
-                                  SiteStatus.vac_statuses[:both_full_time_and_part_time_vacancies],
-                                  SiteStatus.vac_statuses[:part_time_vacancies],
-                                  SiteStatus.vac_statuses[:full_time_vacancies]
-                                ]
-                              )
-                              .where.not(
-                                "site.address1 IN ('', NULL)
-                                AND site.address2 IN ('', NULL)
-                                AND site.address3 IN ('', NULL)
-                                AND site.town IN ('', NULL)
-                                AND site.address4 IN ('', NULL)
-                                AND site.postcode IN ('', NULL)
-                                AND site.latitude IS NULL
-                                AND site.longitude IS NULL"
-                              )
-  end
-
   def available_placements_count
-    @available_placements_count ||= available_placements.count
+    @available_placements_count ||= site_statuses.size
   end
 
   def visa_sponsorship