--- a/app/services/provider_suggestion.rb
+++ b/app/services/provider_suggestion.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ProviderSuggestion
+  include ActiveModel::Model
+
+  attr_accessor :id, :name, :code, :value
+
+  def ==(other)
+    id == other.id
+  end
+end