--- a/app/services/subject_suggestion.rb
+++ b/app/services/subject_suggestion.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class SubjectSuggestion
+  include ActiveModel::Model
+
+  attr_accessor :name, :value
+end